feat: persist session in cstate
request.session could be used to store some shared state (e.g., in StimulusReflex). We keep it up-to-date by storing its contents in the cstate.

We also wrap the original request.session provider into a proxy class which lazily "unwrap" the stored values (via GlobalID, the same way as with identifiers)